Refactor vimrc into a directory of configs
Parsing out the vimrc file into a directory of config files is helpful
for a few reasons. This makes the plugin list maintainable on one file
rather than having to maintain it on both the main vimrc file and on the
temp file. (Remember the temp file was required to install the plugins
programatically without having plugin settings depend on the plugins
that haven't been installed yet.) Having the general configs that don't
rely on plugins within as separate file is handy when these configs are
loaded into an IDE or other application.